+++ R/nestedcv.R
@@ -34,6 +34,10 @@
 #' @param n_inner_folds Number of inner CV folds
 #' @param outer_folds Optional list containing indices of test folds for outer
 #'   CV. If supplied, `n_outer_folds` is ignored.
+#' @param pass_outer_folds Logical indicating whether the same outer folds are
+#'   used for fitting of the final model when final CV is applied. Note this can
+#'   only be applied when `n_outer_folds` and `n_inner_folds` are the same and
+#'   no balancing is applied.
 #' @param alphaSet Vector of alphas to be tuned
 #' @param min_1se Value from 0 to 1 specifying choice of optimal lambda from
 #'   0=lambda.min to 1=lambda.1se
@@ -164,6 +168,7 @@ nestcv.glmnet <- function(y, x,
                           n_outer_folds = 10,
                           n_inner_folds = 10,
                           outer_folds = NULL,
+                          pass_outer_folds = FALSE,
                           alphaSet = seq(0, 1, 0.1),
                           min_1se = 0,
                           keep = TRUE,
@@ -248,8 +253,18 @@ nestcv.glmnet <- function(y, x,
     
     if (finalCV) {
       # use CV on whole data to finalise parameters
+      foldid <- NULL
+      if (pass_outer_folds) {
+        if (n_outer_folds == n_inner_folds && is.null(balance)) {
+          foldid <- rep(0, length(y))
+          for (i in 1:length(outer_folds)) {
+            foldid[outer_folds[[i]]] <- i
+          }
+        } else message("Cannot pass `outer_folds` to final CV")
+      }
       cvafit <- cva.glmnet(filtx, yfinal, alphaSet = alphaSet, family = family,
-                           weights = weights, penalty.factor = filtpen.factor, ...)
+                           weights = weights, penalty.factor = filtpen.factor, 
+                           nfolds = n_inner_folds, foldid = foldid, ...)
       alphafit <- cvafit$fits[[cvafit$which_alpha]]
       s <- exp((log(alphafit$lambda.min) * (1-min_1se) + log(alphafit$lambda.1se) * min_1se))
       fit <- cvafit$fits[[cvafit$which_alpha]]
@@ -376,6 +391,8 @@ nestcv.glmnetCore <- function(test, y, x, filterFUN, filter_options,
 #' @param y Response vector
 #' @param nfolds Number of folds (default 10)
 #' @param alphaSet Sequence of alpha values to cross-validate
+#' @param foldid Optional vector of values between 1 and `nfolds` identifying
+#'   what fold each observation is in.
 #' @param ... Other arguments passed to [cv.glmnet]
 #' @return Object of S3 class "cva.glmnet", which is a list of the cv.glmnet 
 #' objects for each value of alpha and `alphaSet`.
@@ -391,8 +408,11 @@ nestcv.glmnetCore <- function(test, y, x, filterFUN, filter_options,
 #' @importFrom utils tail
 #' @export
 #' 
-cva.glmnet <- function(x, y, nfolds = 10, alphaSet = seq(0.1, 1, 0.1), ...) {
-  foldid <- sample(rep(seq_len(nfolds), length = length(y)))
+cva.glmnet <- function(x, y, nfolds = 10, alphaSet = seq(0.1, 1, 0.1),
+                       foldid = NULL, ...) {
+  if (is.null(foldid)) {
+    foldid <- sample(rep(seq_len(nfolds), length = length(y)))
+  }
   fit1 <- cv.glmnet(x = x, y = y, 
                     alpha = tail(alphaSet, 1), foldid = foldid, ...)
   if (length(alphaSet) > 1) {